+++ Magick++/demo/piddle.cpp
@@ -40,7 +40,8 @@ int main( int /*argc*/, char ** argv)
     // Start drawing by pushing a drawing context with specified
     // viewbox size
     drawList.push_back(DrawablePushGraphicContext());
-    drawList.push_back(DrawableViewbox(0,0,image.columns(),image.rows()));
+    drawList.push_back(DrawableViewbox(0,0,(ssize_t) image.columns(),
+      (ssize_t) image.rows()));
 
     //
     // Draw blue grid
